requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
azure-storage-blob==12.24.0
Brotli==1.1.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==43.0.1
cssselect2==0.7.0
Deprecated==1.2.14
dj-database-url==2.1.0
Django==5.0.9
django-cors-headers==4.6.0
django-oauth-toolkit==1.4.0
django-redis==5.4.0
djangorestframework==3.14.0
fcm-django==2.0.0
firebase-admin==6.3.0
fonttools==4.47.0
future==0.18.3
geographiclib==2.0
geopy==2.4.1
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
html5lib==1.1
httplib2==0.22.0
idna==3.7
jwcrypto==1.5.6
msgpack==1.0.7
numpy==1.26.2
oauthlib==3.2.0
pandas==2.1.4
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.1
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydyf==0.8.0
pyfcm==1.5.4
PyJWT==2.8.0
pyparsing==3.1.1
pyphen==0.14.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
requests==2.32.0
rets==1.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
weasyprint==60.2
webencodings==0.5.1
wrapt==1.16.0
xmltodict==0.13.0
zopfli==0.2.3
